consolidate config broker and config source
