+++ src/python/sigopt_config/broker.py
@@ -8,31 +8,14 @@ import os
 import json_merge_patch
 import yaml
 
-from sigopt_config.source import ConfigBrokerSource
 from sigopt_config.utils import is_mapping
 
 
-_NO_DEFAULT = object()
-
-
 class ConfigBroker(object):
-  def __init__(self, source):
-    assert isinstance(source, ConfigBrokerSource), error_message
-    self.source = source
-    self.impl = ConfigBrokerImpl(self.source)
-
-  def get(self, name, default=None):
-    return self.impl.get(name, default)
-
-  def get_object(self, name, default=None):
-    return self.impl.get_object(name, default)
-
   @classmethod
   def from_configs(cls, configs):
     merged_config = functools.reduce(json_merge_patch.merge, reversed(list(configs)), {})
-    source = ConfigBrokerSource(merged_config)
-    broker = cls(source)
-    return broker
+    return cls(merged_config)
 
   @classmethod
   def from_directory(cls, dirname):
@@ -42,25 +25,41 @@ class ConfigBroker(object):
         configs.append(yaml.safe_load(config_fp))
     return cls.from_configs(configs)
 
+  def __init__(self, data):
+    self.data = data
+
+  def _split_name(self, name):
+    return name.split(".")
+
+  def __contains__(self, name):
+    try:
+      self[name]
+      return True
+    except KeyError:
+      return False
+
+  def get(self, name, default=None):
+    try:
+      return self[name]
+    except KeyError:
+      return default
+
+  def get_object(self, name, default=None):
+    return self.get(name, default)
+
   def __getitem__(self, name):
-    ret = self.impl.get(name, _NO_DEFAULT)
-    if ret == _NO_DEFAULT:
-      raise KeyError(name)
-    return ret
+    base_dict = self.data
+    parts = self._split_name(name)
 
+    for p in parts[:-1]:
+      base_dict = base_dict.get(p)
+      if not is_mapping(base_dict):
+        raise KeyError(name)
 
-class ConfigBrokerImpl(object):
-  def __init__(self, source):
-    self.source = source
+    if is_mapping(base_dict):
+      try:
+        return base_dict[parts[-1]]
+      except KeyError as ke:
+        raise KeyError(name) from ke
 
-  def get(self, name, default):
-    value = self.source.get(name, default)
-    self._ensure_safe_return(value)
-    return value
-
-  def get_object(self, name, default):
-    return self.source.get(name, default)
-
-  def _ensure_safe_return(self, val):
-    if is_mapping(val):
-      raise Exception("Possibly unsafe .get of JSON object, values might be missing. Please use .get_object instead")
+    raise KeyError(name)
